requirements.txt
Flask==2.1.1
requests==2.27.1
Werkzeug==2.0.3
python-telegram-bot